+++ internal/transactions/handlers/response/response.go
@@ -4,5 +4,5 @@ type NewTransactionResponse struct {
 	TransactionId   string  `json:"transaction_id,omitempty"`
 	AccountId       string  `json:"account_id,omitempty"`
 	OperationTypeId int     `json:"operation_type_id,omitempty"`
-	Amount          float32 `json:"amount,omitempty"`
+	Amount          float64 `json:"amount,omitempty"`
 }
